Critter moving refactroing and other fixes (#86)
* Disabled separation of walk/run on engine level

* AS object info tuning

* Dir changing

* Effects usage simplifying

* Allow null entity in getter (for protos)

* NullGetterForProto for control virtual getter for proto

* Mouse event fix

* Mouse pos refactoring

* Texture color check fix

* Fixed some NotImplementedException

* Game.GetDistance fix

* Exception reporting tuning - message boxes only in client/mapper/editor

* Stack trace separated on verb (for logs) and brief (for message box)

* Contour fixes

* Contour fixes

* Properties dict fix

* Removed exception if texture not found

* Refactoring

* AngelScript exception message

* Dummy pink texture if not found

* Refactoring

* Formatting

* Script Map::RebuildFog

* Removed player disconnection on server fails

* Critter pick fix

* Entities setters tuning for correct field caches

* Fixed crash when windows minimized

* Allow to pass delegates to script calls

* Restored dialog starting

* Some NotImplementedException fixed

* Removed lookup to dict for built-in dialogs

* Dialogs script calls restored + validation during bakering

* Input events tuning

* Common lerp improved

* Server logging with collapsed content

* Compilation fix

* Visual particles tuning

* ModelProjFactor as setting to control models base projection scale

* Static analysis fix

* Stack overflow fix

* Dialogs refactoring

+++ Source/Client/Sprites.cpp
@@ -288,7 +288,7 @@ auto Sprites::RootSprite() -> Sprite*
     return _rootSprite;
 }
 
-auto Sprites::PutSprite(Sprite* child, DrawOrderType draw_order, ushort hx, ushort hy, int x, int y, int* sx, int* sy, uint id, uint* id_ptr, short* ox, short* oy, uchar* alpha, RenderEffect** effect, bool* callback) -> Sprite&
+auto Sprites::PutSprite(Sprite* child, DrawOrderType draw_order, ushort hx, ushort hy, int x, int y, int* sx, int* sy, uint id, uint* id_ptr, int* ox, int* oy, uchar* alpha, RenderEffect** effect, bool* callback) -> Sprite&
 {
     _spriteCount++;
 
@@ -393,12 +393,12 @@ auto Sprites::PutSprite(Sprite* child, DrawOrderType draw_order, ushort hx, usho
     return *spr;
 }
 
-auto Sprites::AddSprite(DrawOrderType draw_order, ushort hx, ushort hy, int x, int y, int* sx, int* sy, uint id, uint* id_ptr, short* ox, short* oy, uchar* alpha, RenderEffect** effect, bool* callback) -> Sprite&
+auto Sprites::AddSprite(DrawOrderType draw_order, ushort hx, ushort hy, int x, int y, int* sx, int* sy, uint id, uint* id_ptr, int* ox, int* oy, uchar* alpha, RenderEffect** effect, bool* callback) -> Sprite&
 {
     return PutSprite(nullptr, draw_order, hx, hy, x, y, sx, sy, id, id_ptr, ox, oy, alpha, effect, callback);
 }
 
-auto Sprites::InsertSprite(DrawOrderType draw_order, ushort hx, ushort hy, int x, int y, int* sx, int* sy, uint id, uint* id_ptr, short* ox, short* oy, uchar* alpha, RenderEffect** effect, bool* callback) -> Sprite&
+auto Sprites::InsertSprite(DrawOrderType draw_order, ushort hx, ushort hy, int x, int y, int* sx, int* sy, uint id, uint* id_ptr, int* ox, int* oy, uchar* alpha, RenderEffect** effect, bool* callback) -> Sprite&
 {
     // Find place
     uint pos;
